Development: Move external authentication backends (LDAP/SAML2/Passkey) into account
Relocate the LDAP/SAML2/Passkey authentication backends out of core into the
account module, where the rest of the auth code and all their integration tests
already live:
- core/config/{LdapEnabled,Saml2Enabled,PasskeyEnabled,SAML2Configuration,SAML2Properties} -> account/config
- core/service/connectors/SAML2Service, core/service/connectors/ldap/LdapAuthenticationProvider -> account/security
- core/dto/{PasskeyDTO,PasskeyAdminDTO}, core/dto/passkey/* -> account/dto(/passkey)

SecurityConfiguration is unchanged: LdapAuthenticationProvider keeps its
@Component("ldapAuthenticationProvider") bean name (resolved via @Qualifier),
so no edit is needed there. The two passkey exceptions and their FailureAnalyzer
intentionally stay in core/exception (consumed by core ExceptionTranslator and
ConfigurationValidator), avoiding new core->account edges and a spring.factories
change.
