+++ src/main/java/de/tum/cit/aet/artemis/account/security/LdapAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.core.service.connectors.ldap;
+package de.tum.cit.aet.artemis.account.security;
 
 import java.util.HashSet;
 import java.util.Locale;
@@ -18,14 +18,13 @@ import org.springframework.security.ldap.SpringSecurityLdapTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
+import de.tum.cit.aet.artemis.account.config.LdapEnabled;
 import de.tum.cit.aet.artemis.account.domain.User;
 import de.tum.cit.aet.artemis.account.repository.UserRepository;
-import de.tum.cit.aet.artemis.account.security.ArtemisAuthenticationProvider;
 import de.tum.cit.aet.artemis.account.service.ldap.LdapUserDto;
 import de.tum.cit.aet.artemis.account.service.ldap.LdapUserService;
 import de.tum.cit.aet.artemis.account.service.user.AuthorityService;
 import de.tum.cit.aet.artemis.account.service.user.UserCreationService;
-import de.tum.cit.aet.artemis.core.config.LdapEnabled;
 import de.tum.cit.aet.artemis.core.security.SecurityUtils;
 
 /**
