+++ src/main/java/de/tum/cit/aet/artemis/account/service/ldap/LdapUserService.java
@@ -19,9 +19,9 @@ import org.springframework.ldap.query.SearchScope;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
+import de.tum.cit.aet.artemis.account.config.LdapEnabled;
 import de.tum.cit.aet.artemis.account.domain.User;
 import de.tum.cit.aet.artemis.account.repository.ldap.LdapUserRepository;
-import de.tum.cit.aet.artemis.core.config.LdapEnabled;
 
 @Lazy
 @Service
